correcting headers and keep solving the problem

+++ prova_2/fila2.h
@@ -3,7 +3,7 @@
 struct fila{
 	int n;
 	int ini;
-	float vet[N];
+	int vet[N];
 };
 
 typedef struct fila Fila2;
@@ -11,11 +11,11 @@ typedef struct filal FilaL2;
 
 
 Fila2 *fila_cria_vet();
-void fila_insere_vet(Fila2 *f, float v);
-float fila_retira_vet(Fila2 *f);
+void fila_insere_vet(Fila2 *f, int v);
+int fila_retira_vet(Fila2 *f);
 int fila_vazia_vet(Fila2 *f);
 void fila_libera_vet(Fila2 *f);
-void fila_insere_ini_vet(Fila2*f, float v);
-float fila_retira_fim_vet(Fila2 *f);
+void fila_insere_ini_vet(Fila2*f, int v);
+int fila_retira_fim_vet(Fila2 *f);
 void fila_imprime_vet(Fila2*f);
